roody_backend/requirements.txt
aiocoap==0.3
aniso8601==2.0.0
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
psycopg2==2.7.3.2
pyserial==3.4
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
Werkzeug==0.14.1
